+++ NEWS.md
@@ -31,6 +31,7 @@ The following optimisers are removed, as they are no longer supported by Tensorf
 ## Minor
 
 * `greta_sitrep()` now checks for minimum versions of software, instead of exact versions. It requires at least Python version 3.8, TensorFlow 2.8.0, and Tensorflow Probability 0.14.0.
+* slice sampler no longer needs precision = "single" to work.
 
 # greta 0.4.3